Port the x-api endpoints that the CLI uses into the main app
TL;DR: This is not ideal, but it's a step in a slightly better
direction, taken in a way that I think will cause the least amount
of pain in terms of people using the CLI.

The most important thing that this will enable, is a fairly
straight-forward change in the data model that will let us get
completely rid of the "version" tests that some tracks use in their
exercises, where they have a test in the test suite that specifies
a version, and the person solving it needs to add a hard-coded value
in their solution to match that version.

For more about this problem and how we envision solving it, see the
discussion in https://github.com/exercism/discussions/issues/157

---

To make this change, I checked which x-api endpoints the CLI actually talks
to, and then copied these into a new Sinatra app within this repository.

I put it inside of the 'api' directory, but I'm mounting it alongside
it.

I also ported over the tests relevant to these endpoints.

I had to make a dirty hack around some stuff in the approvals library.
Long story short, I wrote approvals as an experiment, made some pretty
bad design choices, but the library has been helpful so I still use it.

In this case I'm working around a singleton that never should have seen
the light of day.

This works with the existing CLI with the following change:

    exercism configure --api=http://exercism.io/xapi

Or, to test locally:

    exercism configure --host=http://localhost:4567 --api=http://localhost:4567/xapi

---

In order to make this change live, I think we could:

1. create an FYI issue in discussions explaining the change
1. update all the documentation about configuring the CLI to point
  to use the new URL
1. create a dismissable alert that we can put on everyone's dashboard
  as well as on their notifications page (if they dismiss it in
  either place it will disappear from both)
1. after a couple of weeks, update the endpoints in x-api to report
  an error (which will be displayed by the CLI) that points to the
  FYI issue, and shows the command to update the configuration.

---

This whole API thing has a long and storied history.

In the earliest days of Exercism, when I didn't realize it would turn
into a real thing, I had all the problem specifications and exercise
implementations living within exercism.io.

Then we started getting more tracks, and it was really painful to couple
the changes in the exercises to the changes in the website. I tried
splitting out all the tracks to their own repositories (a good choice)
and splitting out the API that delivered those exrecises into its own
app (somewhat less of a good choice).

This became painful because it turns out I hadn't found the correct
seams, and needed cross-talk between the two APIs.

Extracting all of the track-related stuff into the Trackler library
was a good choice, which cleaned up some painful cross-talk, but
didn't eliminate it. It also has some deploy-related pains (but those
are unrelated to this pull request).

Anyway—by moving the x-api back into the main website app, we get
rid of the final bit of cross talk.

Eventually I want to rewrite the entire API from scratch taking into
account all the things we've learned in the past four years. I think
we have a much better idea of which abstractions we have and what the
API needs to do.